Ev works

+++ src/main/java/com/first/rednotes/dao/NoteRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface NoteRepo extends CrudRepository<Note, Long> {
+public interface NoteRepo extends CrudRepository<Note, Integer> {
     List<Note> findByHeading(String tag);
     List<Note> findByAuthor(User author);
 }
